refactor: :recycle: Refactor lexer_scanner
norm에 맞게 함수 분리
- 함수 개수 초과로 인해 파일도 분리

closes #33
